% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist1.R
\name{rdistk}
\alias{rdistk}
\title{simulation k fois}
\usage{
rdistk(x, p, k = 100)
}
\arguments{
\item{x}{numeric vector representing the values of the random variable}

\item{p}{numeric vector repesenting the probabilities}

\item{k}{numeric}
}
\description{
simulation k fois
}
