% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist2.R
\name{statdisct}
\alias{statdisct}
\title{statistiques et graphes}
\usage{
statdisct(x, p, k)
}
\arguments{
\item{x}{numeric vector representing the values of the random variable}

\item{p}{numeric vector repesenting the probabilities}

\item{k}{numeric}
}
\description{
statistiques et graphes
}
